requirements.txt
django>=4.2
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
drf-yasg>=1.21
django-filter>=23.5
pytest>=8.0
pytest-django>=4.8
python-dotenv>=1.0
